requirements.txt
aio-tmdb==0.1.3
bugsnag==4.1.1
discord.py==1.7.3
dislash.py==1.4.9